implementation of forward propagation in MLP

+++ multilayer_perceptron/activation_functions/_exponent_helper.py
@@ -2,5 +2,5 @@
 
 class _ExponentHelper:
 
-    def calculate_exponent(self, number: int) -> float:
+    def calculate_exponent(self, number: float) -> float:
         return 2.718 ** number
